Pare down gem requirements to the bare minimum
This is a minimal API app and doesn't need the full weight of Rails
behind it (or indeed some of the default gems that come along for the
ride). Removing them enables smaller deployment sizes and faster setup
for development, as well as a smaller third-party code surface area in
general.

The `tzinfo-data` gem is unnecessary as development on GOV.UK apps will
not happen natively on Windows.