Refactor event pipeline into mini-library
The processing of data from the publishing message queue (and subsequent
pushing to a search engine product) is conceptually almost entirely
separate from the API part of this application. This moves all the
relevant code into a library in `lib/` that can be invoked separately
and doesn't need to form part of the Rails app.

- Move import/event pipeline related code into a semi-detached library
  of its own
- Refactor and rename to make things clearer
- Make message queue name configurable through environment