Make publishing event pipeline more modular
We're not sure right now if the write-side of this project is going to
remain a part of this application (because we may discover that there is
a requirement for shared code between it and the API Rails app), or if
they can safely be separated into completely independent units. Until
then, this makes the pipeline code a bit more modular if we do want to
extract it into another unit.

- Make logger and repository used by `PublishingEventPipeline`
  configurable, and set configuration in Rake task and tests
- Remove dependency on Rails environment from pipeline Rake task
- Consolidate `require` statements for library code into top level
  module file