Decouple repositories from `PublishingEventPipeline`
Now that we have moved the repositories out into their own module, it
doesn't necessarily make sense for them to know about internals of the
pipeline module. If we assume that a repository's knowledge about a
"document"/"indexable thing" can be limited to an ID, some metadata, and
some unstructured content, we can get rid of the `Document` model and
thus a cross-module boundary dependency entirely.

- Remove `PublishingEventPipeline::Document` class
- Make repositories take explicit bare arguments for metadata and
  content