Add JSON schema validation to integration tests
As Discovery Engine won't actually validate schemas for us on ingestion,
we need to be a little bit more careful around our metadata to avoid
accidentally modifying the schema of a datastore. This adds a middle
ground solution of validating schema compliance as part of our
integration tests.

Unfortunately the source of truth for the schema is in a different
repository, and it's still actively "under construction", so to avoid
vendoring schema that's at risk of becoming stale, for now we'll fetch
the schema on running the tests.