Fix over-eager document type ignorelist and overrides
The `String#match?` method will implicitly convert its argument to a
regex, which means that the exact strings in the ignorelist and
ignorelist overrides suddenly become a bit too eager and start
ignoring/overriding non-exact matches. Use `===` (despite Rubocop's
protestations) for the desired behaviour for both strings and regular
expressions.