Add boosting for news content at query time
News content (specifically, content with a content_purpose_supergroup of
`news_and_communications`) doesn't age very well in terms of its
usefulness to mainstream search scenarios. Early on in its life, it may
be more relevant to users but the model hasn't learned that yet due to a
lack of events. Later on, it may have been viewed a lot but no longer be
relevant and potentially misleading to users who don't notice the date.

- Add two promotion and two demotion timescales/cutoffs and apply
  dynamic query time boost specs depending on the content's
  public_timestamp
- Add `timecop` for time freezing in tests (and enforce `safe_mode`)