Opportunistically delete ignored documents
The ignorelist contents and logic may change over time, and we want to
delete any documents that may exist remotely but are now covered by the
ignorelist.

We've set deletion up as an idempotent operation anyway, so there is no
harm in firing off an opportunistic delete request when a document is
ignored.

- Attempt to delete documents remotely that come through as ignored
- Add an example ignored message to integration tests
- Make documents not existing on deletion an `info` rather than `warn`
  log message, and remove TODO comment now that our approach is final
- Fix incorrectly named mock object