Always set document `url` field to production GOV.UK
This field isn't used by the frontend, but rather serves as a key
property in the Discovery Engine schema to allow it to apply "web
signals" to influence ranking. We want it to be the production URL even
in non-production environments to ensure these environments also pick up
those signals.

- Hardcode production URL prefix for document metadata
- Remove unnecessary `plek` gem