Turn `attachments` into `parts`
When a document has "attachments" in Publishing API, these should come
through in search as if they were parts (even though they're technically
not). This is a bit of a fudge from v1, but we have no choice but to
replicate it to keep compatibility and desired functionality.

Unfortunately the message from Publishing API doesn't include the
content of the attachment, so depending on whether we consider that
necessary, we may have to go out to Publishing API and fetch content to
populate the `body` field of generated attachment parts.