Un-ignorelist HTML publications
We've been ignoring these so far in favour of just syncing their parent
document. Following a discussion today, we want to try and have them as
part of search after all, as they contain more useful content than their
parents, and we may be able to pick up valuable web signals this way.