Reject timestamp filters that are the wrong way around
Finder Frontend will happily let users enter a "from" timestamp filter
that is after the "to" value. This doesn't make sense, and Discovery
Engine rightly complains and refuses to service the request.

Until this gets fixed upstream (if ever), flip from and to dates so we
at least return some sort of meaningful results.