Ensure incoming filter parameters are valid
Finder Frontend should be validating these, but isn't, and it's easier
to deal with this here. We sometimes get garbled, clearly non-human
search requests into Finder Frontend with nonsensical query
parameters like the following example:

```
GET /search/all?commit=Refresh+results&page=3&
    topics%5B%5C%5C%5C%5C%5D=all&world_locations%5B%5C%5C%5C%5C%5D=all
```

This then gets passed on to us and we end up with a query parameter that
should always be a single value or an array being parsed into a hash,
causing errors.

This commit adds some simple validation to ensure we discard such
invalid parameters if they make it as far as our app.