Fix sync failures due to excessive taxons
There are a few documents that are tagged to several hundred taxons in
the GOV.UK topic taxonomy, for example broad travel advice that is
considered applicable to all countries. This causes issues with an
undocumented (other than in Cloud Retail Search documentation on tags)
limit of 250 elements in an array field in Vertex, leading any document
tagged to more than 250 taxons to fail to sync.

- Truncate `part_of_taxonomy_tree` field to a maximum of 250 array items
- Reorder how we parse the taxonomy tree so that parent taxons come
  first (before direct taxons), as these are needed for faceting so we
  want them to have a better chance of being in the first 250