Flip logic and name of version check method
The `DocumentVersionCache#outdated?` method was named a bit confusingly
as it doesn't just check whether a document is stricly outdated, but
also whether its ID exists in the cache to begin with (and whether the
incoming document has a payload version).

This renames it to a positive `#sync_required?` which better describes
its function, and flips the logic around.