Add service and task to update autocomplete denylist
The data file for the denylist is stored in a GCP Cloud Storage bucket
rather than as configuration in the app itself, as the denylist may
contain sensitive language. The plan is to potentially move this into a
private gem in the future.

In the meantime, this adds a service class to handle the occasionally
required task of purging and re-importing the denylist, as well as a
Rake task to invoke it.