Add feature flag to toggle autocomplete
This adds a feature flag to disable autocomplete in case of unforeseen
problems by returning an empty set of completion results in the
response.

This is controlled by the env var `ENABLE_AUTOCOMPLETE`, and will allow
us to consolidate the logic behind disabling autocomplete in a single
place instead of having it scattered across several apps (as the
component will not render completions if no results are returned).