Add `cdoc_url` to metadata
This is a new field mirroring the document's content ID. We need to have
this as an additional, separate field as VAIS's search quality
evaluation tooling cannot be configured to use our existing content ID
field.

Despite the name, this field does not have to be a URL, and can be any
arbitrary unique identifier. We've decided that since we already use
content ID as a unique ID for documents everywhere, we should also use
it for this use case (instead of the relative link or URL) for
consistency.

see https://github.com/alphagov/govuk-infrastructure/pull/1866
see https://cloud.google.com/generative-ai-app-builder/docs/evaluate-search-quality