Metrics: Remove search and autocomplete counters
These are equivalent to the default `http_requests_total` metric and
don't need to be separate (and also don't follow Prometheus naming
conventions anyway). While the `search_requests` metric theoretically
had some labels we can't infer from HTTP requests alone, we don't need
that data in alerting or dashboards, and can get it from GA4 on the
user-facing side of things if we do want to analyse it.