Metrics: Remove more superfluous metrics
This removes the two metrics `message_processing_errors` and
`discovery_engine_requests`. These are superfluous on the happy path (as
they should be covered by the count of documents we synced), and on the
unhappy path every error is one too many, so instead of tracking it in a
dashboard or alerting through metrics, Sentry should be the primary
place for these to be surfaced, triaged, and handled by the team.