Purge user events at end of retention period
This adds a `DiscoveryEngine::UserEvents::Purge` service that can be
used to purge user events older than a certain date, and a Rake task to
run it.

In a perfect world, we'd be able to tell Discovery Engine to purge all
events older than $END_OF_RETENTION_PERIOD. Alas, we're limited to a
maximum of 30 days of purging at a time, with even shorter windows
encouraged for performance reasons - so we've ended up deciding to purge
a week's worth of events at the tail end of the retention period on a
daily basis (giving us some buffer in case it fails to run on a given
day).