Check that Autocomplete errors are thrown
At the moment all errors reported from Google for autocomplete are being
sent to Sentry. This makes the Sentry reporting quite noisy.

We want to catch and discard the errors that cannot be actioned whilst
still recording the errors that are of use. The first step is to make
sure that all Google Cloud errors are being raised to the caller.