Raise a custom error for Autocomplete errors
Rather than just swallowing internal errors from VAIS that cannot be
actioned, the errors should be caught and an explicit 500 error
returned to the user.

A new custom InternalError class has been created so that the
controller only needs to rescue one error.

In the next commit the controller will rescue this error and display an
appropriate message to the user.

The custom error exists in its only file as currently the only custom
error that exists.