Gracefully recover from RabbitMQ failures
We suffer from occasional, intermittent RabbitMQ errors
e.g. AMQ::Protocol::EmptyResponseError, that cause the Document Sync
Worker task to crash.

Catching this error avoids filling up Sentry with errors that can't be
actioned. The worker is also terminated with exit code 1, and we rely on
Kubernetes to take care of restarting the worker.