Check that Search response errors are thrown
Al the moment all errors reported from Google Discovery Engine for
search are being sent to Sentry. This makes the Sentry reporting quite
noisy.

We want to catch and discard the errors that cannot be actioned whilst
still recording the errors that are of use. The first step is to make
sure that all errors raise by the Discovery Engine are being raised to
the caller.

To make the tests work, `subject!(:result_set) { search.result_set }`
needed to be moved to within the context it is being used. That's
because unlike `subject`, `subject!` evaluates the block before the
specs run and therefore the result of
`client.search(discovery_engine_params).response` was being memoized
before the new tests ran, meaning that an error was never raised.