Catch "InternalError" Search errors
At the moment all errors reported from Google for autocomplete are being
sent to Sentry. This makes the Sentry reporting quite noisy.

Errors like `Google::Cloud::InternalError` are not actionable
therefore they should instead be caught and logged.

An existing custom error is being thrown back to the caller so that an
explicit 500 error can be returned to the user.