Label search requests with consumer details
This adds user labels to search requests based on the user agent string,
allowing us to differentiate between consumers for analytics and billing
purposes.

A `consumer_group` label is set based on the kind of consumer
(`web`/`app`/`other`), and a more specific `consumer` label is set to the
consuming app itself (e.g. `finder-frontend` or `app-ios`).

These labels don't affect results, but can be used in the Google Cloud
billing console to differentiate between different consumers, so costs of
search use can be tracked and associated with specific teams.