Create a query set for evaluations
The v1 beta API includes two services:

- sample_query_set_service
exposes an endpoint to create an empty query set

- sample_query_service
exposes an endpoint to fetch performance data from a big query
table and push that data into the empty query set.

The SampleQuerySet class is hardcoded to fetch data from the clickstream table [1]
which is partioned by month. We will always fetch the previous month as this will
ensure we have a full month of data to evaluate.

[1] https://github.com/alphagov/govuk-infrastructure/blob/74f2bb7b8c6604d247cce7d358b25fe616c60027/terraform/deployments/search-api-v2/automated_evaluation.tf#L124