Fetch evaluation
In subsequent work the results of the evaluation will be sent to Prometheus, but for
now this method just returns the raw results and outputs them to the console.

The get_evaluation endpoint returns an operation object that exposes a state method.
Quality metrics are only returned once the state is :SUCCEEDED[1]

1 = https://cloud.google.com/ruby/docs/reference/google-cloud-discovery_engine-v1beta/latest/Google-Cloud-DiscoveryEngine-V1beta-Evaluation