Use location as top-level resource
Up until now, we've had the default collection passed in through the
app's environment as the top-level resource that other things are nested
underneath. Since we will now have to start dealing with resources that
are on a sibling level to collections (sample query sets), this changes
the app to use the location as the top level resource.

- Introduce `Collection` model to represent collections, and set it as
  the parent for engines and datastores
- Remove configuration of default collection in favour of default
  location