Add a month label to the data for Prometheus
We want to run the daily quality metrics reporting task against
sample query sets for the previous two months. Add the labels :last_month
and :month_before_last so that we can create graph in prometheus to allow
us to compare quality metrics. If we send the month name, the value of the label
will change, so the last month/previous month comparison won't be apparent.