Add `MonthInterval` to represent month
This adds a new `MonthInterval` class that represents the time span of
one calendar month that we want to create sample query sets for (for
example, July 2025).

This is a better representation than a "standard" Ruby date object, as
it purely refers to a given month and year. It also offers a convenient
factory method to get a previous month's period.