Create SampleQuerySets class
SampleQuerySets is responsible for creating a SampleQuerySet object for each
dataform table ID, for the date provided.

This paves the way for adding more id's to the BIGQUERY_TABLE_IDS array in a later commit.

We need to pass a table_id to SampleQuerySet to support this new class.

Update rake task to use the new SampleQuerySets .all method