Use symbols :last_month or :month_before_last to call methods
The only class that really needs access to date logic is the
SampleQuerySet class, because it is in this class that the values
for the partition date are calculated.

This class can automatically calculate the partition date if you pass
it the month labels :last_month or :month_before_last, but it can also accept
a month and a year integer for occasions when a specific sample query set
needs to be created. Ie in the troubleshooting rake task
quality:setup_sample_query_set

In all other classes and rake tasks, a the month label can be used to determine
which quality metrics we want to create, collect or report.