Add the table_id as a label
When we send data to prometheus we provide labels, which we can use
to create dashboards that are helpful. Currently the labels we send
are top-k and month. We also need to add the table_id so that we can
see which metrics are for clickstream data, and which are for binary.

I've called the label "dataset" rather than table_id as it feels a bit
easier to understand.