Add error handling for failed evaluations
Our daily report_quality_metrics rake task should create 4 evaluations:

clickstream - last month and month before last
binary - last month and month before last

The DiscoveryEngine evaluations API works in two steps, first you request an evaluation,
which returns an operation object[1]. And only once the operation state changes to succcess
can we then fetch the results. The create takes a while, and we continue to poll
until we get a success message. We recently failed to create one of the four evaluations, as
GCP returned an AlreadyExistsError[2].

This commit rescues the error, retries three times, and then raises in Sentry so that we know the
evaluation is missing.

[1] = https://cloud.google.com/agentspace/agentspace-enterprise/docs/reference/rest/v1alpha/projects.locations.evaluations/create?hl=en
[2] = https://govuk.sentry.io/issues/6756595901/?referrer=slack&notification_uuid=c6a2d660-26ff-477b-830c-4cb2b95f2e2d&alert_rule_id=14659787&alert_type=issue