Don't try and send empty metrics to Prometheus
The evaluations API recently returned empty quality metrics
due to a misconfiguration of our sample query sets. Adding a guard
clause to prevent the rake task that calls this class from blowing
up.

Alternatively we could send 0 if the dig returns nil, see [1], but
that feels unexpected. Better not to send empty metrics.

[1] https://github.com/prometheus/client_ruby/blob/5c64546870ef306c34374a813c61188fcc1fa546/lib/prometheus/client/gauge.rb#L16C21-L16C36