Remove retry when google returns an error
The retry was added in https://github.com/alphagov/search-api-v2/pull/515

Vertex allows only one evaluation to be run at a time per project. This code
is retrying a creation request when google is already actively working on a
different one.

Evaluations can take upto an hour to complete, so retrying after 3 seconds is just
using up our quota of evaluation requests and is highly unlikely to succeed.

We could set the retry period to 30 minutes, but that feels like a bad plan as
then we're queuing requests on a bit of an opaque schedule.

So it's probably just easier to remove the retry and let the task fail, with documentation to
explain that in this situation the task should be run at a later time. We could also consider
a task that polls vertex to see if there are active evaluations running.