Fetch ListResults for a given evaluation
The GCP EvaluationsService exposes a list_evaluation_results endpoint [1], which returns detailed query
level metrics for a given evaluation. An evaluation must exist for the list results to be fetched.

If none exists when this method is called, we should create one, and if one already exists then we
should not create another.

[1] = https://cloud.google.com/ruby/docs/reference/google-cloud-discovery_engine-v1beta/latest/Google-Cloud-DiscoveryEngine-V1beta-EvaluationService-Client#Google__Cloud__DiscoveryEngine__V1beta__EvaluationService__Client_list_evaluation_results_instance_