Implement EvaluationsRunner class
Create all evaluations, and fetch their list_evaluation_results, which
we are referring to here as detailed metrics.

In a follow up commit, this class will be extended to push the detailed
metrics for a gcp bucket.
In a follow up PR, this class will be extended to fetch the aggregate metrics
as well - replacing the existing Evaluations class[1]
[1] https://github.com/alphagov/search-api-v2/blob/main/app/services/discovery_engine/quality/evaluations.rb