Skip pushing metrics to Prometheus when rake tasks are run locally
This is because the Prometheus push gateway is local to the cluster in integration,
staging or production, so it doesn't make sense and is not easily possible to push
metrics to a particular environment. If metrics need to be pushed to Prometheus,
the task should be run in the relevant cluster.