Move registry creation to Metrics Evaluation
Each new instance of PrometheusReporter creates a new instance of
MetricsEvaluation. This is causing an AlreadyRegisteredError error to be
thrown:

Prometheus::Client::Registry::AlreadyRegisteredError
search_api_v2_evaluation_monitoring_recall has already been registered (Prometheus::Client::Registry::AlreadyRegisteredError)

It looks as though the registry gauges can only be created once, and
then have values added to it. A way to make sure that the same instance
of MetricsEvaluation is used for all instances of PrometheusReporter by
making MetricsEvaluation a singleton.

You can't easily initialise a Singleton with a variable, so instead the
registry which was being passed to the MetricsEvaluation is now being
set up there instead.