Make Metrics::Evaluation a singleton class
Each new instance of PrometheusReporter creates a new instance of
MetricsEvaluation. This is causing an AlreadyRegisteredError error to be
thrown:

Prometheus::Client::Registry::AlreadyRegisteredError
search_api_v2_evaluation_monitoring_recall has already been registered (Prometheus::Client::Registry::AlreadyRegisteredError)

It looks as though the registry gauges can only be created once, and
then have values added to it. A way to make sure that the same instance
of MetricsEvaluation is used for all instances of PrometheusReporter by
making MetricsEvaluation a singleton.

Including the [Singleton] module makes the "new" method private. Instead
to instantiate the class the first time, and refer to the same instance,
an `instance` method is made available.

[Singleton]: https://medium.com/@traveling-coder-abhishek/singleton-pattern-explained-from-ruby-on-rails-point-of-view-150e5ee61d47