Add serving_config_name private method to evaluation
This is being read from the api response of GCP's
"get_evaluation" method on the EvaluationService Client,
to give us the ability to trace back which serving configuration
an evaluation was actually run with. Reading the value
directly from the response was thought to be more reliable
than reusing an attribute for the value that was used to
create an evaluation.