Add a 10second wait between evaluation requests
The evaluations API can only handle one evaluation create request at a time.
Whilst an evaluation is being created, the api returns a Google::Cloud::AlreadyExistsError if
another create request is received.

Our code in the Evaluation class checks that the evaluation has a state of :SUCCEEDED before
a fresh evaluation request is sent. However we are still seeing this error returned from the client.

Adding a sleep of 10 seconds is a pragmatic attempt to work around this, given our limited visibility
of what is causing the AlreadyExistsError to be returned.