Check for PENDING evaluations
The evaluations service can only create one evaluation at a time. Sometimes
when we send a create request, the api still has an evaluation in a PENDING state, despite
our checks. This commit is an attempt to guard against this, by fetching all evaluations[1]
and if any have a state of pending we wait for it to complete.

[1]https://cloud.google.com/ruby/docs/reference/google-cloud-discovery_engine-v1beta/latest/Google-Cloud-DiscoveryEngine-V1beta-EvaluationService-Client#Google__Cloud__DiscoveryEngine__V1beta__EvaluationService__Client_list_evaluations_instance_