Ignore FAILED evaluations
Evaluations can be in one of five states when returned by the evaluations API
STATE_UNSPECIFIED, PENDING, RUNNING, SUCCEEDED, FAILED
https://cloud.google.com/ruby/docs/reference/google-cloud-discovery_engine-v1beta/latest/Google-Cloud-DiscoveryEngine-V1beta-Evaluation-State

We don't want to wait for a FAILED evaluation, so we should break our while loop
if a pending or running evaluation fails instead of succeeds.

We don't need to raise an error at this point, because this check is just to clear the way for us to create
a new evaluation. If a new evaluation fails, we should think about how we'll handle it but that functionality
is not being added here.