Update month returned by partition date
Updates the partition date that we use to create a sample query set

From: 1st day of the previous month
To: 1st day of the current month

This is so that a sample query set created on the first of November for
example, will import judgement list data from August, September and
October. Instead of what we are currently doing which is importing
judgement list data from July, August and September.

Also updates the sample query set tests to reflect the changes to the
partition date. It was easier to freeze the date one month earlier so
that the partition date matches the "current" month, rather than
amending all of the expectations.

The labels will updated to better reflect the month of the partition
date in a later change.