Add a wait and retry to evaluation create requests
When we request the creation of a new evaluation, we are intermittently
receiving a Google::Cloud::AlreadyExistsError. This is despite us checking
for running evaluations before submitting a fresh create request.

This commit adds error handling and retries at the point at which the error is raised:

If the create request returns an Google::Cloud::AlreadyExistsError
sleep for 60seconds and retry a maximum of 5 times before logging
the error.