Remove v1beta client methods
The evaluations rely on features in the
google-cloud-discovery_engine-v1beta gem. This gem is incompatible with
the main client (google-cloud-discovery_engine) gem, so the evaluations
code has been moved to it's [own app](https://github.com/alphagov/search-api-v2-beta-features).

`cloud_storage_service` has also been removed as this was only used to
send evaluation metrics to a GCP bucket.